requirements.txt
# DOTENV
python-dotenv
# DATABASE
psycopg2
sqlalchemy
# AGENT
langchain
langchain_groq